Allow a way to specify a custom pixel tolerance in a ref test https://bugs.webkit.org/show_bug.cgi?id=149828
Reviewed by Martin Robinson.

Tools:

When determining if a ref test failed, look for <meta name=fuzzy> in the test file,
and if present, compare the actual pixel differences with the range allowed by the
fuzzy metadata. This may cause a failing reference to match, or a passing reference
(based on ImageDiff internal tolerance) to fail.

In the case of mismatch ref tests, the fuzziness is taken into account to determine
if the mismatch still occurs. If not, then the mismatch test fails.

* Scripts/webkitpy/layout_tests/controllers/single_test_runner.py:
(SingleTestRunner._fuzzy_tolerance_for_reference):
(SingleTestRunner._compare_output_with_reference):

LayoutTests:

Add some ref tests using <meta name=fuzzy> to test fuzzy-based passes and failures.

* fast/harness/image-diff/fuzz-2-25-expected.html: Added.
* fast/harness/image-diff/fuzz-2-25.html: Added.
* fast/harness/image-diff/fuzz-2-36-over-expected-mismatch.html: Added.
* fast/harness/image-diff/fuzz-2-36-over.html: Added.
* fast/harness/image-diff/fuzz-2-36-under-expected-mismatch.html: Added.
* fast/harness/image-diff/fuzz-2-36-under.html: Added.
* fast/harness/image-diff/zero-fuzz-expected.html: Added.
* fast/harness/image-diff/zero-fuzz.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@287323 268f45cc-cd09-0410-ab3c-d52691b4dbfc