Convert css3/scroll-snap/scroll-snap-wheel-event.html to use monitorWheelEvents() https://bugs.webkit.org/show_bug.cgi?id=234526
Reviewed by Wenson Hsieh.

Source/WebCore:

This test sends a stateless wheel event and waits for the timer and subsequent scroll snap
animation, which is already tracked by WheelEventTestMonitor, but the code had a bug where
the monitor wasn't informed if no animation was required.

* platform/mac/ScrollingEffectsController.mm:
(WebCore::ScrollingEffectsController::statelessSnapTransitionTimerFired):

LayoutTests:

Convert the test to use a UIHelper function which makes use of WheelEventTestMonitor.

* css3/scroll-snap/scroll-snap-wheel-event.html:
* resources/ui-helper.js:
(window.UIHelper.async statelessMouseWheelScrollAt):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@287376 268f45cc-cd09-0410-ab3c-d52691b4dbfc