[GTK][a11y] Test /webkit/WebKitAccessibility/selection/listbox is flaky with ATSPI https://bugs.webkit.org/show_bug.cgi?id=234665
Reviewed by Adrian Perez de Castro.

The problem is that selection-changed can be called multiple times for list boxes. When the notification is
deferred by the AXObject cache, it's ensured to be called only one for every element in the following run loop
iteration, but when not deferred the signal is emitted for every notification.

* accessibility/atspi/AccessibilityObjectAtspi.h:
* accessibility/atspi/AccessibilityObjectSelectionAtspi.cpp:
(WebCore::AccessibilityObjectAtspi::selectionChanged): Ensure we don't emit the selection-changed signal more
than once in the same run loop iteration.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@287455 268f45cc-cd09-0410-ab3c-d52691b4dbfc