Include a few more tag names to search when running modal container detection https://bugs.webkit.org/show_bug.cgi?id=234652
Reviewed by Darin Adler.

Source/WebCore:

Broaden the list of tag names that we heuristically scan for the search term. This particular modal container
detection failure was due to the search term appearing inside a `b` element, though I'm adding a few additional
types just to be safe. Since we're checking more than just few tag names, we also take this opportunity to
convert this into a compact AtomString hashtable of tag names that we should scan in the process of detecting
modal containers.

Test: ModalContainerObservation.DetectSearchTermInBoldTag

* page/ModalContainerObserver.cpp:
(WebCore::matchesSearchTerm):

Tools:

Add an API test that's representative of the website where modal container observation failed (due to the search
term being inside a `b` element).

* TestWebKitAPI/Tests/WebKitCocoa/ModalContainerObservation.mm:
(TestWebKitAPI::TEST):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@287497 268f45cc-cd09-0410-ab3c-d52691b4dbfc