Access key should work on focusable element. https://bugs.webkit.org/show_bug.cgi?id=72359
Reviewed by Darin Adler.

Source/WebCore:

Elements that have "accesskey" attribute and are focusable should be focused when you use the access key,
this is happening for some elements but not all.
To fix the issue this patch moves accessKeyAction() implementation from some subclasses to HTMLElement,
which makes us match other browsers.

This patch changes the behavior for HTMLAnchorElement, as it was not focused before, but that matches other browsers too.

Test: fast/dom/accesskey-focus-element.html

* html/HTMLAnchorElement.cpp:
(WebCore::HTMLAnchorElement::accessKeyAction): Deleted.
* html/HTMLAnchorElement.h:
* html/HTMLButtonElement.cpp:
(WebCore::HTMLButtonElement::accessKeyAction): Deleted.
* html/HTMLButtonElement.h:
* html/HTMLElement.cpp:
(WebCore::HTMLElement::accessKeyAction):
* html/HTMLSelectElement.cpp:
(WebCore::HTMLSelectElement::accessKeyAction): Deleted.
* html/HTMLSelectElement.h:

LayoutTests:

Add new test to check the behavior of accesskey on focusable and not focusable DIV, together with other elements
like INPUT, SELECT and A.
Update expectations for fast/forms/access-key.html test, as now more elements are focused than before.
Apart from that we had to modify accessibility/mac/search-predicate-visited-links.html as that was pressing a link
and the link is now focused.

This will make us also pass css/selectors/focus-visible-024.html from WPT, but that test cannot run properly on WebKit
due to webkit.org/b/234139.

* accessibility/mac/search-predicate-visited-links.html: Update test due to change of behavior.
* fast/dom/accesskey-focus-element-expected.txt: Added.
* fast/dom/accesskey-focus-element.html: Added.
* fast/forms/access-key-expected.txt: Update expectations.
* platform/ios/TestExpectations: Skip new accesskey test.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@287529 268f45cc-cd09-0410-ab3c-d52691b4dbfc