Undownloaded iCloud Photos are inserted as broken images when attachment element is enabled https://bugs.webkit.org/show_bug.cgi?id=234803 rdar://82318259
Reviewed by Darin Adler.

Source/WebCore:

Currently, when inserting file paths via paste or drop that correspond to undownloaded files in iCloud Drive,
we show a generic empty file as the file preview, with "Zero Bytes" as the attachment subtitle when the
attachment element is enabled (i.e., Mail compose when using WebKit2). Even worse, for undownloaded image files,
we'll attempt to insert them as broken images.

Mail handles this scenario by detecting that we've dropped an undownloaded attachment, downloads the attachment,
and eventually updates the attachment element with the downloaded data. In the meantime, however, it's not ideal
for WebKit to represent the attachment that is still being downloaded as a broken image or empty file.

To address this, make a slight adjustment to detect that the file we're inserting is not locally available and
render it as an attachment element with progress="0". The client is then expected to update this progress as the
download is taking place, and update the attachment data once it's complete.

Test: WKAttachmentTestsMac.InsertNonExistentImageFileAsAttachment

* editing/cocoa/WebContentReaderCocoa.mm:
(WebCore::attachmentForFilePath):

Tools:

Add a new API test to exercise the change. See WebCore/ChangeLog for more information.

* TestWebKitAPI/Tests/WebKitCocoa/WKAttachmentTests.mm:
(nonexistentFilePath):
(TestWebKitAPI::TEST):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@287547 268f45cc-cd09-0410-ab3c-d52691b4dbfc