MediaRecorder should support the bitsPerSecond option https://bugs.webkit.org/show_bug.cgi?id=234784
Reviewed by Darin Adler.

LayoutTests/imported/w3c:

* web-platform-tests/mediacapture-record/MediaRecorder-bitrate.https-expected.txt:

Source/WebCore:

In case bitsPerSecond is set, we now compute audioBitsPerSecond and videoBitsPerSecond value according to it.
We set audioBitsPerSecond to a tenth of bitsPerSecond and set videoBitsPerSecond to the remaining available bandwidth.
We then use minimum values of 8kbps for audio and 80kbps for video.
As per spec, we compute these values at creation, start and stop times.
We introduce dedicated class members for audio and video bitrates, similarly to the spec defining corresponding slots.

Test: http/wpt/mediarecorder/MediaRecorder-bitrate.html

* Modules/mediarecorder/MediaRecorder.cpp:
* Modules/mediarecorder/MediaRecorder.h:
* platform/mediarecorder/MediaRecorderPrivate.cpp:
* platform/mediarecorder/MediaRecorderPrivate.h:
* platform/mediarecorder/MediaRecorderPrivateAVFImpl.cpp:

LayoutTests:

* http/wpt/mediarecorder/MediaRecorder-bitrate-expected.txt: Added.
* http/wpt/mediarecorder/MediaRecorder-bitrate.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@287613 268f45cc-cd09-0410-ab3c-d52691b4dbfc