Modal container observer should detect and suppress elements that prevent user interaction https://bugs.webkit.org/show_bug.cgi?id=234695
Reviewed by Tim Horton.

Source/WebCore:

Add support for additionally detecting and hiding viewport-constrained elements whose sole (apparent) purpose
is to prevent the user from interacting with the rest of the page, after suppressing a detected modal
container. See comments below for more details.

Test: ModalContainerObservation.HideUserInteractionBlockingElement

* page/ModalContainerObserver.cpp:
(WebCore::ModalContainerObserver::setContainer):

When setting (and hiding) a detected modal container, queue an internal task to try and identify an element
underneath the modal container that may be preventing user interaction with the rest of the page (note that this
must be queued as an async task because it may trigger layout, unlike `updateModalContainerIfNeeded`). See below
for more information.

(WebCore::ModalContainerObserver::hideUserInteractionBlockingElementIfNeeded):

Implement a simple heuristic to detect elements that block user interaction after the modal container has been
hidden. To do this, we hit-test five locations relative to the absolute rect for fixed-position elements: one
hit-test in the center of the rect, and 4 more near all of the corners (shifted by 1px towards the center).

If all 5 hit-tests find the same viewport-constrained element with no rendered children that isn't also the
document or body element, we flag it as the `m_userInteractionBlockingElement` and hide it during style
adjustment in the same way as the modal container (see below).

(WebCore::ModalContainerObserver::revealModalContainer):
* page/ModalContainerObserver.h:
(WebCore::ModalContainerObserver::shouldHide const):

Add `m_userInteractionBlockingElement` and hide both the modal container as well as this new element during
style adjustment.

Tools:

Add an API test (as well as a new test page) to exercise the fix.

* TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj:
* TestWebKitAPI/Tests/WebKit/modal-container-with-overlay.html: Added.
* TestWebKitAPI/Tests/WebKitCocoa/ModalContainerObservation.mm:
(TestWebKitAPI::TEST):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@287671 268f45cc-cd09-0410-ab3c-d52691b4dbfc