WorkerMessagePortChannelProvider::takeAllMessagesForPort should guarantee execution of the takeMessagePort callback https://bugs.webkit.org/show_bug.cgi?id=234883 <rdar://86708232>
Reviewed by Alex Christensen.

In case we enqueue a worker task, and before the task is executed, the worker is terminated,
the task will not be executed and will be destroyed in worker thread.
Instead, we should make sure to execute the callback in main thread.
Introduce MainThreadCompletionHandler for that purpose.
Covered by existing tests.

* dom/messageports/WorkerMessagePortChannelProvider.cpp:
(WebCore::MainThreadCompletionHandler::MainThreadCompletionHandler):
(WebCore::MainThreadCompletionHandler::~MainThreadCompletionHandler):
(WebCore::MainThreadCompletionHandler::complete):
(WebCore::WorkerMessagePortChannelProvider::takeAllMessagesForPort):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@287679 268f45cc-cd09-0410-ab3c-d52691b4dbfc