Ensure that the top layer is always anchored to the RenderView https://bugs.webkit.org/show_bug.cgi?id=231292 <rdar://problem/83941854>
Reviewed by Simon Fraser.

Source/WebCore:

A fixed positioned block level box is not necessarily anchored to the RenderView (see canContainFixedPositionObjects).
This patch ensures that we skip those ancestors and return the RenderView as the containing block for the top layer content.

https://fullscreen.spec.whatwg.org/#new-stacking-layer
"If the top layer position property computes to fixed, its containing block is the viewport, and the initial containing block otherwise."

* rendering/RenderObject.cpp:
(WebCore::RenderObject::containingBlock const):
(WebCore::containerForElement):

LayoutTests:

* TestExpectations: offsetFromAncestorContainer expects no transformed renderer in-between the
dialog and the RenderView's layer. It will go away with the layer re-parenting work.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@287683 268f45cc-cd09-0410-ab3c-d52691b4dbfc