REGRESSION(r285618):A crash may happen when calculating the color-interpolation of a referenced SVG filter https://bugs.webkit.org/show_bug.cgi?id=234918 rdar://86928631
Reviewed by Simon Fraser.

Source/WebCore:

When building a referenced SVGFilter, the color-interpolation property
of the effect element is needed. If the effect element does not have a
renderer, we fallback to the computed style property value. And if the
SVG filter is inside an <iframe> which has media queries, a Document::
updateLayout() will be forced. Building the SVGFilter should not invoke
an updateLayout() since this may not be safe and out of order.

Test: css3/filters/reference-filter-color-interpolation-update-layout.html

* svg/graphics/filters/SVGFilterBuilder.cpp:
(WebCore::colorInterpolationForElement):

LayoutTests:

* css3/filters/reference-filter-color-interpolation-update-layout-expected.txt: Added.
* css3/filters/reference-filter-color-interpolation-update-layout.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@287710 268f45cc-cd09-0410-ab3c-d52691b4dbfc