A gentle momentum scroll into an edge sometimes fails to rubberband, stopping abruptly https://bugs.webkit.org/show_bug.cgi?id=234896
Reviewed by Tim Horton.

During the momentum phase of a scroll gesture we can hit the edge and try to determine that
we should start a rubberband animation. If, at this time, we happen to be at the edge, then
we'd fail to start the animation and the momentum would appear to stop abruptly.

Two fixes were needed. First,
ScrollingEffectsController::startRubberBandAnimationIfNecessary() looks at whether we're
currently stretched, but we might be passing through the edge on the way into a rubberband,
so also need to check for a stretch force.

Second, ScrollAnimationRubberBand needs to not stop prematurely if, in its first callback,
it's still at the edge even though it's animating into a rubberband. The easy fix here is to
just avoid completion during the first few frames.

I was unable to make a reliable layout test for this.

* platform/mac/ScrollAnimationRubberBand.mm:
(WebCore::ScrollAnimationRubberBand::serviceAnimation):
* platform/mac/ScrollingEffectsController.mm:
(WebCore::ScrollingEffectsController::startRubberBandAnimationIfNecessary):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@287714 268f45cc-cd09-0410-ab3c-d52691b4dbfc